requirements.txt
Django~=2.0.6
django-taggit==1.0.0
Pillow
social-auth-app-django
